Refactor Illumina simulator into modular helpers
