+++ tests/test_illumina_insilicoseq_channel.py
@@ -1,12 +1,12 @@
 
 from genecoder.simulators.illumina import IlluminaInSilicoSeqChannel
-import genecoder.simulators.illumina as illumina
+import genecoder.simulators.illumina.cli as illumina_cli
 import genecoder.random_utils as random_utils
 
 
 def test_fallback_deterministic(monkeypatch):
     monkeypatch.setenv("GENECODER_SIM_SEED", "1")
-    monkeypatch.setattr(illumina.shutil, "which", lambda _: None)
+    monkeypatch.setattr(illumina_cli.shutil, "which", lambda _: None)
     random_utils._RNG = None
     channel = IlluminaInSilicoSeqChannel(error_rate=0.2)
     first = channel.simulate("ACGTACGTACGT")
@@ -18,12 +18,12 @@ def test_fallback_deterministic(monkeypatch):
 
 def test_cli_invocation(monkeypatch):
     called = []
-    monkeypatch.setattr(illumina.shutil, "which", lambda _: "/usr/bin/insilicoseq")
-    monkeypatch.setattr(illumina, "_parse_env_options", lambda _: [])
+    monkeypatch.setattr(illumina_cli.shutil, "which", lambda _: "/usr/bin/insilicoseq")
+    monkeypatch.setattr(illumina_cli, "_parse_env_options", lambda _: [])
     def fake_run(cmd, seq):
         called.append((cmd, seq))
         return "ok"
-    monkeypatch.setattr(illumina, "_run_external", fake_run)
+    monkeypatch.setattr(illumina_cli, "_run_external", fake_run)
     ch = IlluminaInSilicoSeqChannel(error_rate=0.1)
     result = ch.simulate("ACGT")
     assert result == "ok"
